Workaround a weird `mypy` error about type mismatch in grpcio
Without the `type: ignore`, this triggers an error:

> Argument "trailing_metadata" to "AioRpcError" has incompatible type
> "tuple[_Metadatum, ...]"; expected "Metadata"

My guess is this is a bug in grpcio bug, the documentation[1] says both
the constructor and the method should return `Metadata`. Checking the
code, it also have the correct type hinting, but `mypy` shows another
error in `trailing_metadata()` saying it is overriding the parent class
incorrectly. Also the type hints are correct but type checking is
completely ignored, as for example internally in the error class the
trailing_metada is `Metadata | None`, but then it is returned as just
`Metadata` without checking for `None` or explicitly `cast`ing. This is
why I suspect it is a grpcio bug, but I couldn't find any reports for
it. It might be a good idea to open one at some point if necessary.

I'm not sure why this is not happening in the SDK.

[1] https://grpc.github.io/grpc/python/grpc_asyncio.html#grpc.aio.AioRpcError

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>