Unify the importing of protobuf generated files
We group all the necessary disabling of `pylint` checks and make the
names more clear regarding if they are from the generated protobuf
files or not.

This also reenable the checks after they are not needed anymore, so the
linter can still check the rest of the file for other (valid) errors.

We also need to use `.ValueType` instead of `.V` because `pylint` also
think that `V` doesn't exist. It seems more clear to use `.ValueType`
anyway.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>