Improve `TypeVar` for `ComponentData`
It is simpler and more future proof to use a *bound* `TypeVar` instead
of specifying the `ComponentData` subclasses one by one as a
*constraint*. With this we also make sure that the resolved type is the
actual subclass of `ComponentData`.

The comment is also removed as it doesn't really provide much useful
information, all the information is either present in the `TypeVar`
declaration and documentation or is repeating the documentation for
`_component_data_task` and it is outdated because it is used in other
methods too.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>