Raise a new `ClientError` instead of `AioRpcError`
We want to avoid exposing the underlying details of the client, so in
the future we can change the libraries used to connect to the service,
or even the protocol.

This also removes the re-building of the exceptions, which was pretty
unnecessary and needed some `type: ignore`s due to typing issues in
`grpcio`.

We should probably create a separate error for timeouts in the future.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>