Fix call to `list_connections`
Since betterproto use type hints, it expects the `ConnectionFilter`
`starts` and `end` to be `list`s.

We might also just take a `list` in the `connections()` methods instead
to avoid the copy, maybe ensuring that the filter elements are unique
(which was probably the reason to make them a `set`) is not that
important.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>