Stop using the _grpc_server context manager
In future commits we'll remove the testing grpc server in favour of
just mocking the stub, so we don't need to really listen to a port and
include in the unit tests all the gRPC machinery, when we only want to
test the client does what it is supposed to do when getting some data
from the stub.

This also removes another level of extra indentation.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>