Properly await for `AddInclusionBounds`
Without this, the function call will never actually run. I'm not sure if
this was working anyways because of some obscure implementation detail
in grpcio, but in normal python a not awaited async function should
never actually run.

Besides this, for sure errors can't be really caught without the await
either.

This was detected while migrating to betterproto, as betterproto has
correct type hints.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>