Make the client accept a server URL to connect to
This URL is a string with the form
`grpc://hostname[:<port:int=9090>][?ssl=<ssl:bool=false>]`, meaning that
the `port` and `ssl` are optional and default to 9090 and `false`
respectively.

This makes it more convenient to take server URLs from config files and
enviroment variables, but it also means gRPC channels are not exposed
directly to users anymore, so the internal implementation to connect to
the gRPC server can be changed without it being a breaking change (we
could potentially even change the protocol completely to use something
else rather than gRPC).

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>