Implement `GetMicrogridMetadata`
Implement the `GetMicrogridMetadata` RPC call in the
`MicrogridApiClient`. To do this we also need to implement the data
structures that will be used to wrap the data returned by the API.

These wrappers use a more Pythonic approach than the protobuf messages.
For example, if some fields don't come in the serialized data, a `None`
will be used instead of using the protobuf defaults. This is to avoid
ending up exposing wrong data (for example, if we don't have location
information, it doesn't make sense to tell the user we are located at
latitude 0 and longitude 0.

Most wrappers should probably be moved to `api-common` in the future.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>