Configure dependabot to have separate minor updates for in-devel deps
The `frequenz-client-base` and `frequenz-microgrid-betterproto`
dependencies are in-development and every minor update is potentially
(and very likely) a breaking change. Because of this we want dependabot
to create separate PRs for minor updates for these dependencies.

For patch updates we still want to group them together, because they not
supposed to be breaking changes.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>