Revert "Migrate the client code to use betterproto"
This partially reverts commits:

* Migrate the client code to use betterproto
  (e1594173754e74ae7df6153006f2a8a91eeeddf9)
* Fix component tests to use betterproto
  (b5b52d5ba8a986a643b763137ea10f5b747522f5)
* Fix client test to use betterproto
  (b3efe71c15e42a51610181ff7292e3c5f69b0f64)

It preserves the changes to tests to use mocks instead of a fake but
full gRPC server, and adapts the new code to use the grpcio library and
the generated code from the traditional google protobuf compiler.

Since the `client-base` library currently being used doesn't support
grpcio, we also need to update that library too, but we do that in
a separate commit.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>