Implement `ListComponents`
Implement the `ListComponents` RPC call in the `MicrogridApiClient`. To
do this we also need to implement the data structures that will be used
to wrap the data returned by the API.

These wrappers use a more Pythonic approach than the protobuf messages,
and translate most enums that encode types of components to concrete
Python classes. This way, we can use the type system to ensure we are
using the right type of component in the right place, and enables using
match statements to handle different types of components in an
exhaustive way.

Most wrappers should probably be moved to `api-common` in the future.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>